Notion: set_state
Factor out updating the row by a hashmap of props, and reuse for this
operation.

This isn't used yet; it should be called by sync according to release
date details.

Also add a useful query and a dir to stick some future queries in for
manually poking around in the db