Refactor notifications
Collect notifications into a notifier struct and hold it in Sync.
Instead of printing in place, accumulate notifications here and run them
at the end.

This was a bit fiddly as a lot of things suddenly needed mutable borrows
within Sync as they have to update the state of the notifier, and thus
the state of self.