Add clap
Implement subcommands using an approach demonstrated in wasmer:

https://github.com/wasmerio/wasmer/blob/
0c94019d746880daac2f92a2d198ee677ee1e14/lib/cli/src/cli.rs#L39

Add a stub import-wishlist and a sync command for the main operation.

Refactor db migration into somewhere it can be reused by any subcommands
which need to ensure the db is migrated (probably most of them)

Connect functionality also refactored to there for ease of db
connection. Might need a bit more sophisticated handling in future but
for now continuing to panic on connection failure is ok.