Refactor notifications
The model of having a notifier attached to the sync struct was
overcomplicated; it makes more sense to classify these as SyncEvents,
significant events which happened during syncing, and collect them as
the result of running the sync tasks -- which currently returned no
result anyway.

The results are then prepared as a SyncReport by the cli-level Sync
handler.

One bit of smell here was I had to make the repo field public on Sync to
allow it to be moved back out of the Sync once it was done, and used by
the report next. This probably implies I should just give it a mutable
borrow instead of moving it, but that may require some refactoring.