Configurable hosts etc, another step for it test
Inject in steam + notion URLs to allow a wiremock test approach to work
in future.

Ran into a small snag here which will make testing more difficult
though: the notion client API is still going to go to api.notion.com. In
theory I don't need to test that at the network level anyway, I
shouldn't test the notion client.

However, I have a lot of logic in the TryFrom<Page> for GameNote, so if
I simply abstract over the client and mock that abstraction I won't be
testing any of that logic -- and I can't provide the notion client types
in the test because they're private. This makes testing that layer
difficult.

As things stand I may as well just mock out the whole NotionGamesRepo in
the test because it's effectively already my abstraction over the notion
client, with the exception of the operation I had to do manually with
ureq (which I could separate out into a different struct)