Adapt to changes in wishlist steam API
The wishlist.json previously relied on has changed to a complex
streaming protobuf format which is going to be a much more difficult
source of the data, so fall back on using userdata JSON data from
another URL.

This source doesn't contain "added to wishlist" dates or other data,
just app IDs forming the current wishlist, so I'll have to approximate
the added date as when the import was completed.

Update instructions on where to source this data.