Use official wishlist API
It turns out the previous two wishlist implementations were largely
redundant because there is, in fact, an official IWishlistService API
which isn't documented by steam, and can be used to get the data more
officially.

Rip out the import-wishlist command and revert parts of the last commit,
reinstating the WishlistedGame model and using it to insert wishlist
data into the db.

Add additional models + logic to fetch wishlist data as part of the
regular daily scrape.