Add a diagnostic query for missing release dates
This seems to happen occasionally and often coincides with steam
reporting a game is released, with no release date string at all, even
though it's not actually released.

This will summarise games we know about which have null or blank release
date string, and check if they're showing as released or not as well.